Move vendor styles to their own folder
