+++ wdn/templates_6.0/js-src/plugins/other/datatables.js
@@ -1,4 +1,4 @@
-import datatablesCssUrl from '@scss/components-js/_datatables-wdn.scss?url';
+import datatablesCssUrl from '@scss/components-js/_datatables.scss?url';
 import { loadStyleSheet } from '@js-src/lib/unl-utility.js';
 
 // Storing the state whether the plugin is initialized or not
