+++ wdn/templates_6.0/js-src/plugins/other/jquery-ui.js
@@ -1,4 +1,4 @@
-import jqueryUiCssUrl from '@scss/components-js/_jquery-ui-wdn.scss?url';
+import jqueryUiCssUrl from '@scss/components-js/_jquery-ui.scss?url';
 import { loadStyleSheet } from '@js-src/lib/unl-utility.js';
 
 // Storing the state whether the plugin is initialized or not
